code/backend/requirements.txt
azure-functions~=1.20.0
azure-functions-durable~=1.2.9
azure-identity~=1.16.0
azure-storage-blob~=12.19.1
langchain~=0.1.16
langchain-openai~=0.1.3
pydantic-settings~=2.2.1
pydantic~=2.7.0
